stress-mergesort: cast return from stress_mwc32 to int32_t
Explicitly use type cast to clean up implicit cast warning

+++ stress-mergesort.c
@@ -156,7 +156,7 @@ static int stress_mergesort(const stress_args_t *args)
 
 	/* This is expensive, do it once */
 	for (ptr = data, i = 0; i < n; i++)
-		*ptr++ = stress_mwc32();
+		*ptr++ = (int32_t)stress_mwc32();
 
 	stress_set_proc_state(args->name, STRESS_STATE_RUN);
